requirements.txt
APScheduler==3.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
decorator==5.1.1
environs==9.5.0
flake8==6.1.0
idna==3.4
marshmallow==3.20.1
mccabe==0.7.0
numpy==1.26.1
packaging==23.2
pandas==2.1.2
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tornado==6.3.3
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
